Move averageMeasurements out of comp for testing
